ipc: semop's nsops is the width the kernel takes it at
sys_semop and sys_semtimedop declare nsops `unsigned`, so the syscall
entry narrows the register and acts on the low word. Passed through at 64
bits it was compared against SEMOPM whole: a count above 2^32 answered
E2BIG where the kernel does the operations the low word asks for, and one
that is only high bits answered E2BIG where the kernel answers EINVAL for
a count of zero.

Every other argument in this family is already narrowed to its prototype
(semget's nsems, every id and flag word); nsops was the one left wide.
Nothing on AArch64 reaches it by accident, since a 32-bit register write
zero-extends — it takes a caller genuinely passing a 64-bit value — but
the narrowing is what the kernel does and costs a cast.

+++ src/monitor/sysvipc.c
@@ -495,10 +495,14 @@ long cng_sysvipc_handle(long nr, long a0, long a1, long a2, long a3, long a4) {
     switch (nr) {
     case __NR_semget:
         return do_semget((s32)a0, (long)(s32)a1, (s32)a2);
+    /* nsops is `unsigned` in both kernel prototypes, so the top half of the
+     * register is discarded rather than folded into the count — a distinction
+     * only a caller passing a 64-bit value can see, and there the kernel takes
+     * the low word and proceeds where an unnarrowed one answers E2BIG. */
     case __NR_semop:
-        return do_semop((s32)a0, (const void *)a1, (u64)a2, -1);
+        return do_semop((s32)a0, (const void *)a1, (u32)a2, -1);
     case __NR_semtimedop:
-        return do_semtimedop((s32)a0, (const void *)a1, (u64)a2,
+        return do_semtimedop((s32)a0, (const void *)a1, (u32)a2,
                              (const void *)a3);
     case __NR_semctl:
         return do_semctl((s32)a0, (s32)a1, (int)a2, (u64)a3);
